requirements.txt
pandas
numpy
faker
psycopg2-binary
matplotlib
seaborn
ipykernel
